+++ test/jest/jest-transformer.js
@@ -1,0 +1,6 @@
+const babelJest = require('babel-jest');
+
+const babelConfig = require('./babel.config');
+
+module.exports = babelJest.createTransformer(babelConfig);
+